Moved Scroll method to webUtil.js
Moved Scroll method to webUtil.js
removed helpers.js
