+++ tests/milo/failedblock.test.js
@@ -1,15 +1,25 @@
+/* eslint-disable import/named */
+/* eslint-disable import/extensions */
 /* eslint-disable no-restricted-syntax */
 /* eslint-disable no-await-in-loop */
+import { WebUtil } from '../../libs/webutil';
+
 const { expect, test } = require('@playwright/test');
 const parse = require('../../libs/parse.js');
 const failedBlock = require('../../features/milo/failedblock.spec.js');
 const selectors = require('../../selectors/milo/failedblock.selectors.js');
-const helpers = require('../../libs/helpers.js');
 
 // Parse the feature file into something flat that can be tested separately
 const { name, features } = parse(failedBlock);
 
+// Create test global page/util variables
+let webUtil;
+
 test.describe(`${name}`, () => {
+  // before each test block
+  test.beforeEach(async ({ page }) => {
+    webUtil = new WebUtil(page);
+  });
   features.forEach((props) => {
     test(props.title, async ({ page, browser }) => {
       await page.goto(props.url);
@@ -18,8 +28,8 @@ test.describe(`${name}`, () => {
 
       // Added scrolling for failed block JS to load.
       // Without it, test provides false count for validation checking.
-      await page.evaluate(helpers.scroll, { direction: 'down', speed: 'slow' });
-      await page.evaluate(helpers.scroll, { direction: 'up', speed: 'fast' });
+      await webUtil.scrollPage('down', 'slow');
+      await webUtil.scrollPage('up', 'fast');
 
       const locator = page.locator(selectors[props.tag]);
       const count = await locator.count();
